move parse_args to helpers
